refactor(op-conductor-ops): convert to installable Python package
Move modules into op_conductor_ops/ package directory, enable
package-mode in pyproject.toml, add poetry script entry point,
fix imports to use relative notation, and add missing rich
dependency declaration.

This enables building the tool as a wheel for Melange APK packaging.
