+++ op-conductor-ops/op_conductor_ops/config.py
@@ -1,5 +1,5 @@
-from network import Network
-from sequencer import Sequencer
+from .network import Network
+from .sequencer import Sequencer
 import toml
